Move getComputedMatrix to fix circular import
